# HR-forge-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Vhark",
    "description": "OpenApi documentation for Hr-forge-api",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Vhark",
      "email": "contact@vhark.com"
    },
    "license": {
      "name": "No license"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://some-url.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/employee/{id-or-email}": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Find employee by ID or email",
        "description": "Endpoint for retrieving an employee by ID or email address.",
        "operationId": "getEmployee",
        "parameters": [
          {
            "name": "id-or-email",
            "in": "path",
            "description": "Employee ID or email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employee"
        ],
        "summary": "Update employee details",
        "description": "Endpoint for updating employee details by ID or email.",
        "operationId": "updateEmployee",
        "parameters": [
          {
            "name": "id-or-email",
            "in": "path",
            "description": "Employee ID or email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employee"
        ],
        "summary": "Delete employee",
        "description": "Endpoint for deleting an employee by ID or email.",
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id-or-email",
            "in": "path",
            "description": "Employee ID or email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/employee/password/{id}": {
      "put": {
        "tags": [
          "Employee"
        ],
        "summary": "Update employee password",
        "description": "Endpoint for updating an employee's password by ID (admin operation).",
        "operationId": "updateEmployeePassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Employee ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/employee": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Find all employees",
        "description": "Endpoint for retrieving a paginated list of all employees.",
        "operationId": "getAllEmployees",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "firstName"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Sort direction (ASC or DESC)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ASC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEmployeeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/employee/search": {
      "get": {
        "tags": [
          "Employee"
        ],
        "summary": "Search employees by name or surname",
        "description": "Endpoint for searching employees by part of their first or last name with pagination and sorting.",
        "operationId": "searchEmployees",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Part of first or last name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "John"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "firstName"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Sort direction (ASC or DESC)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ASC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseEmployeeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmployeeRequest": {
        "required": [
          "birthDate",
          "departmentName",
          "email",
          "firstName",
          "hireDate",
          "lastName",
          "positionName",
          "salary"
        ],
        "type": "object",
        "properties": {
          "departmentName": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "hireDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AdminPasswordRequest": {
        "required": [
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "newPassword": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string"
          }
        }
      },
...
      "EmployeeResponse": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          },
          "positionName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "hireDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageResponseEmployeeResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}